Update links.sh for use without libtflbp-sh
Also added a safety check before beginning `rm` & `ln` operation.

+++ devutils/links.sh
@@ -1,18 +1,18 @@
 #!/bin/sh
 
-#----------------------------------------------------------------------------------
+#------------------------------------------------------------------------------
 # Project Name      - i3Config/devutils/links.sh
 # Started On        - Sun 22 Oct 00:15:02 BST 2017
-# Last Change       - Tue 10 Dec 00:11:50 GMT 2019
+# Last Change       - Wed 11 Nov 16:42:25 GMT 2020
 # Author E-Mail     - terminalforlife@yahoo.com
 # Author GitHub     - https://github.com/terminalforlife
-#----------------------------------------------------------------------------------
+#------------------------------------------------------------------------------
 # Just a simple, quick script to update the hard links when changing branches.
-#------------------------------------------------------------------------------MAIN
-
-. /usr/lib/tflbp-sh/ChkDep
-
-ChkDep ln rm mkdir
+#
+# Dependencies:
+#
+#   coreutils (>= 8.25-2)
+#------------------------------------------------------------------------------
 
 cd "$HOME/GitHub/terminalforlife/Personal/i3Config"
 
@@ -22,9 +22,10 @@ cd "$HOME/GitHub/terminalforlife/Personal/i3Config"
 
 	[ -d "$HOME/.i3a" ] || mkdir "$HOME/.i3a"
 
-	cd source/.i3a
-	for CurFile in *; do
-		rm -v "$HOME/.i3a/$CurFile"
-		ln -v "$CurFile" "$HOME/.i3a/$CurFile"
-	done
+	if cd source/.i3a; then
+		for CurFile in *; do
+			rm -v "$HOME/.i3a/$CurFile"
+			ln -v "$CurFile" "$HOME/.i3a/$CurFile"
+		done
+	fi
 } 2> /dev/null
